Extract rust_admin db-tests harness into tests/db_tests.rs
rust_admin becomes lib + thin bin (main.rs calls rust_admin::run) so
integration tests can drive the real router, config, and schema. The
~700-line SQLite harness moves from routes.rs to tests/db_tests.rs,
wrapped in mod db_tests so the existing Makefile/CI invocation
(cargo test -p rust_admin --features db-tests db_tests) still matches
every test. Visibility bumps limited to the items the harness uses
(router, Config, AppState, AdminMetrics, AntdRestClient, ensure_schema,
status/job constants).
