+++ crates/rust_admin/src/config.rs
@@ -34,61 +34,61 @@ impl AuthCookieSameSite {
 }
 
 #[derive(Clone)]
-pub(crate) struct Config {
-    pub(crate) db_path: PathBuf,
-    pub(crate) antd_url: String,
-    pub(crate) antd_internal_token: Option<String>,
-    pub(crate) antd_payment_mode: String,
-    pub(crate) antd_metadata_payment_mode: String,
-    pub(crate) admin_username: String,
-    pub(crate) admin_password: String,
-    pub(crate) admin_auth_secret: String,
-    pub(crate) admin_auth_ttl_hours: i64,
-    pub(crate) admin_auth_cookie_secure: bool,
-    pub(crate) catalog_state_path: PathBuf,
-    pub(crate) catalog_bootstrap_address: Option<String>,
-    pub(crate) all_catalog_bootstrap_address: Option<String>,
-    pub(crate) cors_allowed_origins: Vec<HeaderValue>,
-    pub(crate) bind_addr: SocketAddr,
-    pub(crate) admin_db_min_connections: u32,
-    pub(crate) admin_db_max_connections: u32,
-    pub(crate) admin_db_connect_timeout_seconds: f64,
-    pub(crate) admin_request_timeout_seconds: f64,
-    pub(crate) admin_upload_request_timeout_seconds: f64,
-    pub(crate) admin_shutdown_grace_seconds: f64,
-    pub(crate) upload_temp_dir: PathBuf,
-    pub(crate) upload_max_file_bytes: u64,
-    pub(crate) upload_min_free_bytes: u64,
-    pub(crate) upload_max_concurrent_saves: usize,
-    pub(crate) upload_read_idle_timeout_seconds: f64,
-    pub(crate) upload_ffprobe_timeout_seconds: f64,
-    pub(crate) ffmpeg_bin: String,
-    pub(crate) ffprobe_bin: String,
-    pub(crate) hls_segment_duration: f64,
-    pub(crate) ffmpeg_threads: usize,
-    pub(crate) ffmpeg_filter_threads: usize,
-    pub(crate) ffmpeg_max_parallel_renditions: usize,
-    pub(crate) ffmpeg_rendition_timeout_seconds: f64,
-    pub(crate) upload_max_duration_seconds: f64,
-    pub(crate) upload_max_source_pixels: i64,
-    pub(crate) upload_max_source_long_edge: i64,
-    pub(crate) upload_quote_transcoded_overhead: f64,
-    pub(crate) upload_quote_max_sample_bytes: usize,
-    pub(crate) final_quote_approval_ttl_seconds: i64,
-    pub(crate) approval_cleanup_interval_seconds: u64,
-    pub(crate) antd_upload_verify: bool,
-    pub(crate) antd_upload_retries: usize,
-    pub(crate) antd_upload_timeout_seconds: f64,
-    pub(crate) antd_quote_concurrency: usize,
-    pub(crate) antd_upload_concurrency: usize,
-    pub(crate) antd_approve_on_startup: bool,
-    pub(crate) antd_require_cost_ready: bool,
-    pub(crate) antd_direct_upload_max_bytes: usize,
-    pub(crate) admin_job_workers: usize,
-    pub(crate) admin_job_poll_interval_seconds: u64,
-    pub(crate) admin_job_lease_seconds: i64,
-    pub(crate) admin_job_max_attempts: i32,
-    pub(crate) catalog_publish_job_max_attempts: i32,
+pub struct Config {
+    pub db_path: PathBuf,
+    pub antd_url: String,
+    pub antd_internal_token: Option<String>,
+    pub antd_payment_mode: String,
+    pub antd_metadata_payment_mode: String,
+    pub admin_username: String,
+    pub admin_password: String,
+    pub admin_auth_secret: String,
+    pub admin_auth_ttl_hours: i64,
+    pub admin_auth_cookie_secure: bool,
+    pub catalog_state_path: PathBuf,
+    pub catalog_bootstrap_address: Option<String>,
+    pub all_catalog_bootstrap_address: Option<String>,
+    pub cors_allowed_origins: Vec<HeaderValue>,
+    pub bind_addr: SocketAddr,
+    pub admin_db_min_connections: u32,
+    pub admin_db_max_connections: u32,
+    pub admin_db_connect_timeout_seconds: f64,
+    pub admin_request_timeout_seconds: f64,
+    pub admin_upload_request_timeout_seconds: f64,
+    pub admin_shutdown_grace_seconds: f64,
+    pub upload_temp_dir: PathBuf,
+    pub upload_max_file_bytes: u64,
+    pub upload_min_free_bytes: u64,
+    pub upload_max_concurrent_saves: usize,
+    pub upload_read_idle_timeout_seconds: f64,
+    pub upload_ffprobe_timeout_seconds: f64,
+    pub ffmpeg_bin: String,
+    pub ffprobe_bin: String,
+    pub hls_segment_duration: f64,
+    pub ffmpeg_threads: usize,
+    pub ffmpeg_filter_threads: usize,
+    pub ffmpeg_max_parallel_renditions: usize,
+    pub ffmpeg_rendition_timeout_seconds: f64,
+    pub upload_max_duration_seconds: f64,
+    pub upload_max_source_pixels: i64,
+    pub upload_max_source_long_edge: i64,
+    pub upload_quote_transcoded_overhead: f64,
+    pub upload_quote_max_sample_bytes: usize,
+    pub final_quote_approval_ttl_seconds: i64,
+    pub approval_cleanup_interval_seconds: u64,
+    pub antd_upload_verify: bool,
+    pub antd_upload_retries: usize,
+    pub antd_upload_timeout_seconds: f64,
+    pub antd_quote_concurrency: usize,
+    pub antd_upload_concurrency: usize,
+    pub antd_approve_on_startup: bool,
+    pub antd_require_cost_ready: bool,
+    pub antd_direct_upload_max_bytes: usize,
+    pub admin_job_workers: usize,
+    pub admin_job_poll_interval_seconds: u64,
+    pub admin_job_lease_seconds: i64,
+    pub admin_job_max_attempts: i32,
+    pub catalog_publish_job_max_attempts: i32,
 }
 
 impl Config {
